+++ tests/builders/test_formex4_builder.py
@@ -155,8 +155,8 @@ def test_build_tbl():
 
     table = Table()
     table.rows[1].nature = "head"
-    table.rows[1].insert_cell([P(u"Expert group")], nature="head")
-    table.rows[1].insert_cell([P(u"First name and surname of the expert")], nature="head")
+    table.rows[1].insert_cell([P(u"Expert group")])
+    table.rows[1].insert_cell([P(u"First name and surname of the expert")])
     table.rows[2].insert_cell([P(u"Control of infectious diseases")])
     table.rows[2].insert_cell([P(u"Michael Angelo BORG")])
     table.rows[3].insert_cell([P(u"Information society")], height=3)
@@ -236,9 +236,9 @@ def test_build_tbl__with_title():
     table = Table()
     table.rows[1].insert_cell([P(u"1 euro =")], width=3, styles={"align": "center"})
     table.rows[2].nature = "head"
-    table.rows[2].insert_cell([P()], nature="head")
-    table.rows[2].insert_cell([P(u"Currency")], nature="head")
-    table.rows[2].insert_cell([P(u"Exchange rate")], nature="head")
+    table.rows[2].insert_cell([P()])
+    table.rows[2].insert_cell([P(u"Currency")])
+    table.rows[2].insert_cell([P(u"Exchange rate")])
     table.rows[3].insert_cell([P(u"USD")])
     table.rows[3].insert_cell([P(u"US dollar")])
     table.rows[3].insert_cell([P(u"1,0029")])