+++ CHANGELOG.rst
@@ -19,6 +19,9 @@ Fixed
 * Change in the class :class:`~benker.styled.Styled`:
   The default value of the *nature* parameter is ``None`` (instead of "body").
 
+* Change in the methods :meth:`~benker.table.RowView.insert_cell` and :meth:`~benker.table.ColView.insert_cell`
+  The *nature* of a cell is inherited from its parent's row (or column).
+
 
 v0.4.1 (2019-04-24)
 ===================